PD-6967 added work I have done over the past week
