+++ lib/control/validation/valid.js
@@ -16,7 +16,7 @@ exports.isJson = isJson;
 exports.request = function(req, res, next) {
   const failures = [];
 
-  if (!req.body.hasOwnProperty('signiture')) {
+  if (!req.body.hasOwnProperty('signature')) {
     failures.push('Missing S/MIME signature request parameter');
   }
